ASoC: tegra-alt: add tas2552 amplifier check
Machine driver is calling dai_set_sysclk() on tas2552 amplifier whenever
a playback session is begun. But this requires the hardware for amplifier
to be present, which is not the case on xavier interposer board.
is_codec_dummy is not a suitable check here and hence replaced it with the
actual codec-dai name which is passed from dts.

Bug 200345514

Change-Id: I7ca6b956328db48a29773543dea95163bf203391
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1559057
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>