nvdisp: fix var-length array sizeof errors
There are a few places in the IMP flow where the nvdisp
driver is using sizeof() on a variable-length
array. This array keeps track of the number of window
entries per head that need to be allocated for the given
IMP settings. To fix the sparse errors, this change removes
the corresponding memset() calls, and explicitly initializes
each entry in the array.

Bug 200299572

Change-Id: Idacf91b8c43a8f4579e0bd5ac10a86157e572b8b
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1559324
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>