vblk: Fix memory corruption due to probe failure
This change replaces kfree with unreserve_mempool
api to ensure that memory not allocated by the
driver is not freed wrongly.

Without this patch, in case of probe failure vblk
driver was freeing the handle to the mempool provided
by IVC layer (external to the driver), causing the
scenario of two independent contexts:
1) IVC layer already maintaining the handle
2) Any entity which got the same memory area via kmalloc
trying to update the same memory area without
any synchronization.

Bug 1990482

Change-Id: I5079759dfd6ca8a61b37d29eff64ed535d03370d
Reviewed-on: https://git-master.nvidia.com/r/1559718
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>