drivers: camera: Fix Mipi Calibration sequence
MIPI calibration can happen irrespective of
sensor state by keeping CSI CIL in reset
state. This change adds calibration to complete
when the CSI is powered ON and based on the sensor
settings. By this approach the mipi lanes are
calibrated and are ready as soon as sensor is
turned ON.

Bug 200324656

Change-Id: Ib2372b4fabeb7d7d3464e5b17ca2effaffac6eb8
Signed-off-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1560331
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jihoon Bang <jbang@nvidia.com>