dt: t210: soc: ahub modules under aconnect
With kernel-4.9, we have moved to upstreamed driver versions of
agic, adma and aconnect. This requires ahub and adma modules to
be acting as child nodes for aconnect and later the power domain
would be registered for aconnect. The patch is added under the
linux version checks.

Bug 200329917

Change-Id: I8e5fc5faa8b5475b5c7c917462c581bbfd99595d
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1560616
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>