ata: ahci_tegra: switch to BPMP powergate provider
The runtime PM subsystem can also handle turning powergates, or power domains,
on and off as needed. This is handled by the 'generic power domain', or genpd
system. The system ensures that a power domain is powered on when any device
contained to it is powered on, and when all are powered off, the power domain
will be powered off as well.

When using runtime PM and genpd, the driver itself does not need to include
any code related to powergating. The power domain will be powered on and off
as required, as long as the runtime PM functions are used.

Bug 1853511

Change-Id: I30a078f17e54144dffe9fe2b404415693a1affee
Signed-off-by: Preetham Ramchandra <pchandru@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1560681
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>