tegra-se-nvhost: Correct the logic used to decide memory type
AES individual requests input buffer is copied to a big chunk of memory
and the same is used to for SE HW operations.
Allocation of this bug chunk of memory is decided based on the input
size which is 512 bytes. If the request is not 512 bytes, driver allocates
dynamic memory. In the current allocation logic, once dynamic memory
is allocated, driver always tries to allocate dynamic memory
for next set of requests even through all of them are 512 bytes size.

Corrected the logic so that driver allocates dynamic memory
depending on the request size.

Bug 1959035

Change-Id: If0cf40da16045e1f532728c1af8af2d8a8603d00
Signed-off-by: Mallikarjun Kasoju <mkasoju@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1562238
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>