drivers: camera: Fix calibration for CSI2
CSI2 mipi lanes calibration is driven by sensor
state. As soon as the mipi pad is powered on the
sensor is also powered/streamed to reach the standby
state for performing the calibration during the
initialization sequence of CSI2 HW.
Address issues with loadable sensor modules, calibration
will be performed only when the sensor is bound.

Bug 200324656

Change-Id: Ia9b2ec15c48bfe39cb78bd9651c2574705b743a3
Signed-off-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1565678
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Frank Chen <frankc@nvidia.com>
Reviewed-by: Josh Kergan <jkergan@nvidia.com>
Reviewed-by: Jihoon Bang <jbang@nvidia.com>