nvdisp: pick default IMP settings for tegrasim
Some platforms, such as tegrasim, expose themselves as
silicon, but do not correctly model MC with respect
to the max EMC rate and total available ISO bw that
are reported. On tegrasim, both of these parameters
are reported as zero. This change extends the existing
WAR that we have in place so that the display bw driver
picks the first entry in the IMP table by default, which
avoids a NULL dereference later on.

Change-Id: I69a841271e73d08b7435621dc349c82d61f97656
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1568131
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Greg Stiehl <gstiehl@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>