dts: t18x: common: Disable lc898212 focusers
Disable both lc898212 focusers on E3323 camera boards in the DT to
cease kernel_warning_test causing failures in GVS due to error spew. A
new flash target is provided in this changeset to re-enable the
focusers for certain SQA processes.

The lc898212 focusers are known to degrade physically over time,
causing intermittent i2c/driver communication failure.

The problem was identified in Bug 1766340, and the focuser was
disabled (by default) in user-mode (Gerrit 1466051). This change is
necessary, because the focuser is powered on with the sensor, causing
it to occasionally trigger I2C corruption.

The lc898212 focuser is not and will not be exercised in any current
or future GVS testplans.

JIRA:
Jira CRTC-1372

NvBugs:
Bug 1788047
Bug 1963171 (main)
Bug 200304103
Bug 200336519

Change-Id: If14f4aed758e13d1f232c40a265b18a70bfc2fc1
Signed-off-by: Vincent Chung <vincentc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1568960
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jihoon Bang <jbang@nvidia.com>