Bullseye: RTCPU driver mod for code coverage.
Bug 2001937

This change creates a communication channel between the
RTCPU and the kernel driver so that an RTCPU image
built to obtain Bullseye coverage data can return
that data to the CCPLEX and the kernel can make it
available in a file in the debug file system.
If during the boot sync the kernel driver does not
detect an RTCPU firmware built with Bullseye support
all allocated resources will be freed and the driver
will behave as though this change does not exist.

Change-Id: I9b9bca11229489c3ae802bb20bbddac203a3f6dd
Signed-off-by: Matthew Lazarowitz <mlazarowitz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1569666
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>