media: i2c: ov9281: Use frame period for sleep
This only applies to an ov9281 configured as an fsync slave.

We need to sleep at least as long as the frame sampling period after
making the fsync slave start streaming (we are assuming that the fsync
master is already started and is generating the fsin signal).

Previously, we were hardcoding the time to sleep based on a frame rate
of 120fps.  Instead, calculate the frame sampling period based on our
frame length setting.

Bug 1982522

Change-Id: I7d5c99c59e3eaf6a5016a3de02fc89e5b769b0aa
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1570497
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Xia Yang <xiay@nvidia.com>
Reviewed-by: Raymond Poudrier <rapoudrier@nvidia.com>
GVS: Gerrit_Virtual_Submit