drivers: camera: Fix sensor mode reporting in V4L2
V4L2 always reports the size of the sensor controls to be
the max size of 30. This change fixes the dimensions
reported by V4L2 to report the sensor_modes found in the
DT. It also set the dimension to zero on TPG nodes as they
do not have sensor modes. This makes testing using v4l2-ctl
easier.

Bug 200350907

Change-Id: I5e1433973ea6faff5c8cb0bdf668cf4aec23ebda
Signed-off-by: Josh Kergan <jkergan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1572216
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>