video: tegra: host: Properly handle errors in finalize_poweron
A device could fail to poweron for whatever reason. In that case, rather
than falling through and doing the rest of device init, bail out of the
finalize_poweron call and return the error.

Similarly, in prepare_poweroff, only execute the poweroff sequence if the
device in question was already powered on.

Bug 200352108

Change-Id: I95b991f6fb5c74146aec327b2d6c089b1f51e9c0
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1572790
Reviewed-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Shridhar Rasal <srasal@nvidia.com>
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>