clk: tegra: dfll: fix use of tune values
There was some confusion on the use of the tune values. They were defined
in 2 different structs, but only initialzed in one of them. Remove most
of the references in struct tegra_dfll_soc_data and adjust the code
accordingly. Only tune_high_min_millivolts is left, because this can be
adjusted using debugfs. Hence copy the initial value from the cvb table
at boot.

Change-Id: I7e7b4bbc02e586b84982aa11b60c8d758c6c5ba1
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1576305
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>