t19x: dts: Add SOC Timer
This patch adds DT node for SOC timer based on TKE.

The SOC Timer is outside of the CPUs hence it can continue
to work even when CPUs are powered off.

Driver uses shared interrupts 0 to 7, one for
each Carmel CPU.

DONT enable this node in platform DT as Arch timer
should be used which is already enabled.

Bug 2003168

Change-Id: I6fadbb83c4f30eeb744b303fceb317448be38312
Signed-off-by: Rohit Khanna <rokhanna@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1576546
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ellis Roberts <ellisr@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>