misc: nct1008: rework the nfactor calculations
Apply the correction on the final temperature instead of on just the
offsets that are present in the fuses. Fixup the nfactor and offset
calculations to reflect this change. Update the CP2 fuse mask as it
was changed for T19x.

Bug 1974573

Change-Id: I8726ac3ed28e3d3ea2be6cc0af34c4ec0b695572
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1577269
(cherry picked from commit 7dadae2892a1b1540fc0cada65112ed1bfeeba89)
Reviewed-on: https://git-master.nvidia.com/r/1639694
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Navneet Kumar <navneetk@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bo Yan <byan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>