arm64: configs: enable SND_DMAENGINE_PCM
SND_DMAENGINE_PCM is selected by SND_SOC_TEGRA_ALT and forced to take
'm' value. This requires the driver module for it to be loaded.
Since SND_DMAENGINE_PCM is part of the core sound driver in linux, this
need not be 'm'. This patch allows the config to be selected from menu
and required dependency is added for SND_SOC_TEGRA_ALT.

SND_SOC_TEGRA_T210REF_MOBILE_ALT is enabled for L4T as well.

Bug 200346429

Change-Id: I979bea143b013b90736480ef466424468a4e4b98
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1577828
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Dipesh Gandhi <dipeshg@nvidia.com>
Reviewed-by: Simon Je <sje@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>