kconfig: implement overlays
Kernel overlays allow the kernel source tree to be sharded into separate
directories, referred to as overlays, which are combined into a single
kernel image during the build process. The overlays can be used e.g. to
house different sets of IP that are released on different schedules. This
approach doesn't require the base kernel Kconfig files and Makefiles to
manually specify when to reference source from the overlays; rather the
build system automatically combines the overlays.

Bug 1978395

Change-Id: Ic2de5fd3c56a7037dcc6996b990b356050077fe1
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1579977
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>