video: host: pva: Fix selftest issue
Selftest was causing smmu error on running some
test in different order.
Reason looks like the ucode starts the test before
the kernel complete the test memory allocation.
Fix added to asking the ucode to wait if it is
running the selftest code until the kernel passes the
run state.

Added print for CCQ8 register following the test done,
that helps to identify the failing test in case of a test
failure.

JIRA PVA-692

Change-Id: I5f17f3511428830f48a9554768f0906f43038941
Signed-off-by: Vinod G <vinodg@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1580152
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-by: Colin Tracey <ctracey@nvidia.com>
GVS: Gerrit_Virtual_Submit