soc: tegra: Generate error if OS macro  is not defined
It is required to have the different DT bindings for
some of modules across kernel version specially Linux
kernel version. The DT binding of modules are different
for K4.4 and K4.9.

The DTS files are also used by non-Linux OS like QNX etc.

To have the DT binding independent of OS and support multiple
version of same OS and multiple OS, add DT binding version
file which defines the module DT binding version.

It is mandatory requirement to pass the OS macros when
compiling the DTS. If it is not available then generate
compilation error.

Bug 200344258

Change-Id: I6804ef1e3c572626afde31b0e0774ce3fb5a8b84
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1572476
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Venkata Jagadish <vjagadish@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>