se-elp:Remove DONE check for RNG1 health
Remove checking for DONE bit in
health check for RNG1 as this is not
a compulsory requirement. If this is retained,
issues might arise when secure_mode process is
accessing RNG1 for the first time.
Also add check for RNG1 alarms during
init too though it is checked after
every operation

Bug 200356385

Change-Id: I7f07caf7be9e37834108aad8197f65c112d5c79f
Signed-off-by: Shravani Dingari <shravanid@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1580391
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>