drivers:camera: Update fmt to be sensor specific
V4L2 driver was providing all the possible color_fmts supported
by the VI, when it should have been providing only the values used
by that sensor. This fix updates the driver to check the formats
supported by the VI with the formats supported by the sensor as
described in the sensor_properties.

JIRA CHWI-1379

Change-Id: Ia7f2a7c5080eeb0c65fd10fa16dadb30195a6b29
Signed-off-by: Josh Kergan <jkergan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1613713
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jihoon Bang <jbang@nvidia.com>
Reviewed-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>