misc: nct1008: deprecate nct wake from sleep
Deprecate NCT's ability to wake during suspend. This was added as
part of commit 319155d2ec5a66f6907e148c5b4bdf27d3d46722 to change the
core vmin if the temperature changed while the device was in suspend.
This is no longer a requirement on chips post T210 and is not being
used.

Change-Id: I28c7b3dfc48f6beb34d0b58465dbc6702012340e
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1586408
Reviewed-by: Navneet Kumar <navneetk@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bo Yan <byan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>