tegra: camera: rtcpu: debug: preallocate mem0
Pre-allocate 6MB in mem0 as output buffer for RTCPU functional
tests. The default allocation will not be made if mem0 is
explicitly initialized as an input buffer.

Together with RTCPU carveout support, this change is aimed
at being able to deprecate and remove the rce-rm scratch
memory driver.

Jira CAMC-509

Change-Id: I95b67dd4a93d2c4eb8700913c129b114f0d1247e
Signed-off-by: Mika Liljeberg <mliljeberg@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1588415
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>